+++ jest.config.js
@@ -23,4 +23,6 @@ module.exports = {
   },
   // To help avoid exhausting all the available fds.
   maxWorkers: 4,
+  globalSetup: "./tests/preTest.js",
+  globalTeardown: "./tests/postTest.js",
 };