+++ src/api/aptos.ts
@@ -62,7 +62,8 @@ export class Aptos {
   }
 }
 
-// extends Aptos interface so all the methods and properties from the other classes will be recognized by typescript.
+// extends Aptos interface so all the methods and properties
+// from the other classes will be recognized by typescript.
 export interface Aptos
   extends Account,
     ANS,
@@ -73,8 +74,8 @@ export interface Aptos
     FungibleAsset,
     General,
     Staking,
-    Transaction,
-    ZkID {}
+    ZkID,
+    Omit<Transaction, "build" | "simulate" | "submit"> {}
 
 /**
 In TypeScript, we can’t inherit or extend from more than one class,