+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+- Omit `"build" | "simulate" | "submit"` from `aptos` namespace
+
 # 1.3.0 (2024-01-03)
 
 - [`Breaking`] Capitalize `TransactionPayloadMultiSig` type